requirements.txt
Flask==0.12.4
Flask-Bcrypt==0.7.1
Flask-Login==0.4.0
pymongo==3.7.0
dnspython==1.15.0
pytest-flask==0.10.0
Flask-Cors==3.0.3
flask-jwt-extended==3.7.0
faker==0.8.13
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
